backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
feedparser>=6.0.11
pydantic>=2.0.0
python-dotenv>=1.0.0
